Bug 437166 - CLA plugin fails if projects can't "Push"
Currently the CLA Gerrit plugin uses "Push" permissions to assume a person is a committer on a project. However not all projects want to have "Push" enabled into their repos in order to require code review.  This change uses canSubmit() instead, since that permission will be available to committers in the foreseeable future.