Restore previous semantic for committer check
During migration to permission backend the committer check was changed.
Restore the previous semantic by checking for Submit permission of
refs/heads/* as it was the case prior to migration to permission
backend. Note, that a change in Gerrit core is needed for that. That's
why this change is built against 3.3.0-SNAPSHOT plugin API.

Signed-off-by: David Ostrovsky <david@ostrovsky.org>